# perf: Limit the cache size for `to_datetime`