Fix #15736.

```
s = pl.datetime_range(date(2000, 1, 1), date(2001, 1, 1), '1s', eager=True)
strings = s.dt.strftime('%Y-%m-%dT%H:%M:%S')
```

Without cache:
```
%timeit strings.str.to_datetime(format='%Y-%m-%dT%H:%M:%S', cache=False)
1.47 s ± 71.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

With cache:
```
%timeit strings.str.to_datetime(format='%Y-%m-%dT%H:%M:%S', cache=True)
```
- Before this PR
```
18.2 s ± 790 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
- After this PR
```
1.88 s ± 59.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Consider that the test data are all unique(The cache will never hit) and large, I think the overhead is acceptable.